Move .repeat tests to separate file and delete old test_observalbe_creation.rb.
